expand_symtabs_matching and missing parameter info
FIXME: Needs a new commit log.  The info below is no longer correct.

Without this fix (a fix to this fix that is now squashed), this would
happen:

 (gdb) complete b function(long)
 b (anonymous namespace)::B::function()
 b (anonymous namespace)::B::function() const
 b (anonymous namespace)::function()
 b (anonymous namespace)::function(int, int)
 b B::function()
 b B::function() const
 b Bn::(anonymous namespace)::B::function()
 b Bn::(anonymous namespace)::function()
 b Bn::(anonymous namespace)::function(int, int)
 b Bn::B::function()
 b Bn::function()
 b Bn::function(int, int)
 b Bn::function(long)
 b gdb::function(long)
 (gdb) complete b function(long)
 b function(long)
 (gdb)

I.e., the first time we'd match every overload.  The second time we'd
match symbols only, and thus find the right overload, only.