Make gdb.base/dmsym.exp independent of Ada
The test is using "set language ada" expecting that to cause GDB to do
Ada symbol name matching.  That doesn't work any longer because GDB
now uses the symbol's language to decide which symbol matching
algorithm to use.

Generalize the test a bit to not rely on Ada name matching rules.

Confirmed that by undoing the fix from commit the test still fails.