fix compare-sections
This command is not remote-specific nowadays.  We shouldn't be calling
set_general_process if we're not even talking to a remote target.

It happened to work as is because set_general_process has an early
return in case the remote target does not support multi-process
extensions, which returns false if not connected to a remote target.
Unless you force it on with "set remote multiprocess-feature-packet
on", in which case gdb would crash...

I think we should move this command elsewhere out of remote.c.  Maybe
to exec.c?