fix gdb.threads/no-unwaited-for-left.exp
was looking at state of an exited thread...  revert back to using
ALL_XXX-style macros, but filtered on target.

The normal_stop change is just a fix for a latent bug exposed on the
branch.  In the branch we always switch to no-thread-selected whenever
we start handling an event, and that catches places using
"inferior_ptid/inferior_thread()" referring to previously-selected
threads instead of the event thread.