Fix new gdb.base/quit.exp tests
This fixes the tests by getting rid of the get_remote_state() free
function.  I.e., no more relying on global state in the remote target.

The only places that need to access the current remote target are in
command implementation, and they use get_current_remote_target() now.

This fixes the quit.exp/detach + extra inferior test because that test
ends up closing the remote target files on quit, and that was calling
get_remote_state with the "wrong" inferior.