Make gdb.base/long-inferior-output.exp fail fast
A local change made GDB lose output, which was caught by this
testcase.  While debugging the problem, I found it useful to have the
testcase fail quickly when it noticed output was missing.

Also, tighten the regexes to make sure that we don't get
spurious/repeated/bogus output between each "this is line number ..."
line.