target_ops/C++: linux_trad_target, MIPS and Alpha GNU/Linux
This converts the code in the new linux-nat-trad.h/c files, that was
split off of inf-ptrace.h/c recently, to a C++ base class, and makes
the MIPS and Alpha GNU/Linux ports use it.

Otherwise, it's a straighforward conversion.

I'd checked with a cross compiler that the MIPS port builds.  I don't
have an Alpha compiler handy.