Introduce switch_to_inferior_no_thread
Several places want to switch context to an inferior and its pspace,
while at the same time switch to "no thread selected".  This commit
adds a function that does that, and uses it in a few places.  The
function is made extern because many more uses will be added in follow
up commits.

gdb/ChangeLog:
yyyy-mm-dd  Pedro Alves <palves@redhat.com>

	* inferior.c (switch_to_inferior_no_thread): New function,
	factored out from ...
	(inferior_command): ... here.
	* inferior.h (switch_to_inferior_no_thread): Declare.