Introduce intrusive_list, make thread_info use it
GDB currently has several objects that are put in a singly linked
list, by having the object's type have a "next" pointer directly.  For
example, struct thread_info and struct inferior.  Because these are
simply-linked lists, and we don't keep track of a "tail" pointer, when
we want to append a new element on the list, we need to walk the whole
list to find the current tail.  It would be nice to get rid of that
walk.  Removing elements from such lists also requires a walk
currently, to find the "previous" position relative to the element
being removed.  To eliminate the need for that walk, we could make
those lists double-linked, by adding a "prev" pointer alongside
"next".  It would be nice to avoid the boilerplace associated with
maintaining such a list manually, though.  That is what the new
intrusive_list type addresses.

With an intrusive list, it's also possible to move items out of the
list without destroying them, which is interesting in our case for
example for threads, when we exit them, but can't destroy them
immediately.  We currently keep exited threads on the thread list, but
we could change that which would simplify some things.

Note that with std::list, element removal is O(N).  I.e., with
std::list, we need to walk the list to find the iterator pointing to
the position to remove.  However, we could store a list iterator
inside the object as soon as we put the object in the list, to address
it, because std::list iterators are not invalidated when other
elements are added/removed.  However, if you need to put the same
object in more than one list, then std::list<object> doesn't work.
You need to instead use std::list<object *>, which is less efficient
for requiring extra memory allocations.  For an example of an object
in multiple lists, see the step_over_next/step_over_prev fields in
thread_info:

  /* Step-over chain.  A thread is in the step-over queue if these are
     non-NULL.  If only a single thread is in the chain, then these
     fields point to self.  */
  struct thread_info *step_over_prev = NULL;
  struct thread_info *step_over_next = NULL;

It would also be nice if we could get the advantages of an intrusive
linked list, while avoiding the boilerplate associated with manually
maintaining it.  That is what this new intrusive_list type is about.

intrusive_list's API follows the standard container interface, and
thus std::list's interface.  I actually based the API on Boost's
intrusive list, here:

 https://www.boost.org/doc/libs/1_73_0/doc/html/boost/intrusive/list.html

Actually, I started out by writing a fully-functional intrusive list
with a different interface, and only after did I migrate to Boost's
interface based on that document.  Regardless, the implementation here
is super simple, while Boost's is complicated and intertwined due to a
lot of customization options, which our version doesn't have.

The easiest way to use an intrusive_list is to make the list's element
type inherit from intrusive_node.  This adds a prev/next pointers to
the element type.  However, to support putting the same object in more
than one list, intrusive_list supports putting the "node" info as a
field member, so you can have more than one such nodes, one per list.

In this patch, I'm making thread_info use the member node method, to
show how it works.  The following patch will switch to inheritance
instead, and you'll see how simple that change is.