AMD lane debugging using Intel's lane debugging support
Basically, hook gdbarch_active_lanes_mask with
AMD_DBGAPI_WAVE_INFO_EXEC_MASK.  In Intel's version, this callback
just reads the register that contains the mask out of the regcache.
(which is not that different from what AMD_DBGAPI_WAVE_INFO_EXEC_MASK
does internally, IIUC).

Then, tweak a couple spots in thread.c to get the right architecture
for the gdbarch calls.  They didn't need this because they have two
inferiors, one for the host, and another for the GPU, so the
inferior's gdbarch is OK for them.