gdb: make linux-nat use async event handler
I initially though this patch would be necessary, but perhaps not.
Alternatives are:

- make linux nat use an async signal handler instead, which could be
  marked directly from the sigchld handler, instead of having our own
  pipe in linux-nat.  async signal handlers use a pipe in a same way,
  but only one shared by all handlers.  That pipe's read end would have
  to be exposed for target_ops::async_wait_fd.

- make linux nat implement "has_events" (in the upcoming "consume
  multiple events patch) by just checking if its own pipe is readable.
  That can be implemented using a poll/select with a timeout of 0.  That
  would required the least amount of changes in linux-nat.

I still like using async_event_handler, because it's a simple to
understand flag: when it's set, we (may) have events to report, when
it's not set, we don't.  It's less fiddling than a pipe.  We still need
the pipe to write to from the sigchld handler, but it just becomes a way
to wake up the event loop when it's blocked on poll/select.

--------

The linux_nat target currently interfaces with the event loop using only
a pipe.  Upon receiving a SIGCHLD, this is roughly what happens:

- linux_nat's SIGCHLD handler (sigchld_handler) runs
- it writes a byte to linux_nat_event_pipe
- this wakes up the event loop
- the event loop calls the handler associated with the fd,
  handle_target_event
- handle_target_event calls inferior_event_handler
- that eventually calls linux_nat_target::wait
- linux_nat_target::wait flushes the pipe
- linux_nat_target::wait prepares the event to core
- if linux_nat_target::wait sees that there might be additional events
  to report, it writes a byte to the pipe to restart this sequence

A subsequent patch adds a way for the infrun loop to ask targets if they
have anything to report.  For targets using async_event_handlers, it's
easy: just check the `ready` flag.  This patch makes linux-nat use an
async_event_handler as the way to mark whether it wants its wait method
to be called.

After the patch this is what happens when receiving a SIGCHLD:

- linux_nat's SIGCHLD handler (sigchld_handler) runs
- it writes a byte to linux_nat_event_pipe
- this wakes up the event loop
- the event loop calls the handler associated with the fd,
  handle_target_event
- handle_target_event calls inferior_event_handler
- that eventually calls linux_nat_target::wait
- linux_nat_target::wait flushes the pipe
- linux_nat_target::wait prepares the event to core
- if linux_nat_target::wait sees that there might be additional events
  to report, it writes a byte to the pipe to restart this sequence

Pedro:
  Move a necessary change to linux-nat.c from the next patch into this
  one.

Change-Id: I484380140ae202e68e49a0e4e9a0134690bcd57f