Drop special way of getting inferior context after a Cygwin signal
Simplify Cygwin signal handling by dropping the special way of getting
inferior context after a Cygwin signal.

I think the reason this existed was because previously we were not able
to unwind through the alternate stack used by _sigfe frames, so without
the hint of the "user" code IP, the backtrace from a signal was
unusable.

Now we can unwind through _sigfe frames, drop all this complexity.

(Restoring this specially obtained context to the inferior (as the code
currently does) skips over the actual signal delivery and handling.
Cygwin has carried for a long time a patch which clears the ContextFlags
in the signal context, so we never attempt to restore it to the
inferior, but that interfers with gdb's ability to modify that context
e.g. if it decides it wants to turn on FLAG_TRACE_BIT)

v2:
Update use of DEBUG_EVENTS, changed in 4ef367bf

v3:
Update for windows_per_inferior struct

Change-Id: I214edd5a99fd17c1a31ad18138d4a6cc420225e3