Fix gdbserver handling of just-exited threads (gdb.threads/interrupted-hand-call.exp race)
When testing against gdbserver, gdb.threads/interrupted-hand-call.exp
occasionally fails like this:

  FAIL: gdb.threads/interrupted-hand-call.exp: continue (timeout)

Hacking-in "set debug infrun" logs we see:

   [infrun] fetch_inferior_event: enter
     [infrun] print_target_wait_results: target_wait (-1.0.0 [process -1], status) =
     [infrun] print_target_wait_results:   -1.0.0 [process -1],
     [infrun] print_target_wait_results:   status->kind = no-resumed
     [infrun] handle_inferior_event: status->kind = no-resumed
     [infrun] handle_no_resumed: TARGET_WAITKIND_NO_RESUMED (ignoring: found resumed)
     [infrun] prepare_to_wait: prepare_to_wait
   [infrun] fetch_inferior_event: exit
   FAIL: gdb.threads/interrupted-hand-call.exp: continue (timeout)
   Remote debugging from host ::1, port 35162
   Process /home/pedro/gdb/binutils-gdb/build/gdb/testsuite/outputs/gdb.threads/interrupted-hand-call/interrupted-hand-call created; pid = 3508901

"set debug remote" logs show:

   [remote] Sending packet: $vCont;c:p398f62.-1#50
   [remote] wait: enter
     [remote] Packet received: N
   [remote] wait: exit

And on gdbserver, we see:

  >>>> entering ptid_t linux_process_target::wait_1(ptid_t, target_waitstatus*, target_wait_flags)
  wait_1: [<all threads>]
  LWFE: waitpid(-1, ...) returned 480013, ERRNO-OK
  LLW: waitpid 480013 received 0 (exited)
  LLFE: 480013 exited.
  deleting 480013
  LWFE: waitpid(-1, ...) returned 480011, ERRNO-OK
  LLW: waitpid 480011 received 0 (exited)
  LLFE: 480011 exited.
  deleting 480011
  LWFE: waitpid(-1, ...) returned 480010, ERRNO-OK
  LLW: waitpid 480010 received 0 (exited)
  LLFE: 480010 exited.
  deleting 480010
  LWFE: waitpid(-1, ...) returned 0, ERRNO-OK
  RSRL: resuming stopped-resumed LWP LWP 480009.480012 at 7ffff7f96376: step=0
    continue from pc 0x7ffff7f96376
  Resuming lwp 480012 (continue, signal 0, stop not expected)
  leader_pid=480009, leader_lp!=NULL=1, num_lwps=2, zombie=1
  CZL: Thread group leader 480009 zombie (it exited, or another thread execd).
  deleting 480009
  LLW: exit (no unwaited-for LWP)
  wait_1 ret = null_ptid, TARGET_WAITKIND_NO_RESUMED
  <<<< exiting ptid_t linux_process_target::wait_1(ptid_t, target_waitstatus*, target_wait_flags)
  Writing resume reply for <null thread>:13
  putpkt ("$N#4e"); [noack mode]

Note gdbserver noticed the leader exited, and deleted it.  And then
returned TARGET_WAITKIND_NO_RESUMED.  But a thread had just been
resumed, so how come gdbserver doesn't know there are resumed threads?

The answer is that lwp 480012 exits (because some other thread killed
the whole process) just while gdbserver was resuming it, in
linux_process_target::resume_one_lwp_throw.  That function throws an
error before marking the lwp/thread as not-stopped.

If we try to resume a stopped thread, and we get back an ESRCH, the
best is to just pretend that it resumed successfully, and let the wait
path collect the exit code.  This is what this patch does.

It may be better to remove the whole
linux_process_target::resume_one_lwp try/catch wrapping.  That was
borrowed from gdb, where the ptrace call is done from code shared with
other OS and so we can't give special meaning to ESRCH there without
some more abstraction.  Here we don't have that issue though.